+++ frontend-web/webclient/app/Applications/Jobs/Create.tsx
@@ -52,7 +52,6 @@ import {SshWidget} from "@/Applications/Jobs/Widgets/Ssh";
 import {connectionState} from "@/Providers/ConnectionState";
 import {Feature, hasFeature} from "@/Features";
 import {useUState} from "@/Utilities/UState";
-import {flushSync} from "react-dom";
 import {Spacer} from "@/ui-components/Spacer";
 import {injectStyle} from "@/Unstyled";
 import {UtilityBar} from "@/Navigation/UtilityBar";
@@ -153,7 +152,7 @@ export const Create: React.FunctionComponent = () => {
         numberOfNodes: 1,
         product: null
     });
-    const [reloadHack, setReloadHack] = useState<{ importFrom: Partial<JobSpecification>, count: number } | null>(null);
+    const [reloadHack, setReloadHack] = useState<{importFrom: Partial<JobSpecification>, count: number} | null>(null);
     const [insufficientFunds, setInsufficientFunds] = useState<InsufficientFunds | null>(null);
     const [errors, setErrors] = useState<Record<string, string>>({});
     const [initialSshEnabled, setInitialSshEnabled] = useState<boolean | undefined>(undefined);
@@ -292,6 +291,7 @@ export const Create: React.FunctionComponent = () => {
         return [...injected, ...fromApp, ...workflowInjectedParameters];
     }, [application, injectedParameters, workflowInjectedParameters, estimatedCost]);
 
+
     React.useEffect(() => {
         if (application && provider) {
             const params = parameters.filter(it =>
@@ -354,21 +354,23 @@ export const Create: React.FunctionComponent = () => {
         // Load resources
         // Note(Jonas): An older version could have run with one of these resources while a newer might not allow them.
         // Therefore, check to see if allowed!
+        // Note(Jonas) Pt. II: The actual injection of resources should happen after the function terminates, as React will re-render the component,
+        // and only then will the required input-fields be present. The setTimeout-callback will then be called to fill in the newly created input-fields.
         if (folderResourceAllowed(application)) {
             const newSpace = createSpaceForLoadedResources(folders, resources, "file");
-            injectResources(newSpace, resources, "file");
+            setTimeout(() => injectResources(newSpace, resources, "file"), 0);
         }
         if (peerResourceAllowed(application)) {
             const newSpace = createSpaceForLoadedResources(peers, resources, "peer");
-            injectResources(newSpace, resources, "peer");
+            setTimeout(() => injectResources(newSpace, resources, "peer"), 0);
         }
         if (ingressResourceAllowed(application)) {
             const newSpace = createSpaceForLoadedResources(ingress, resources, "ingress");
-            injectResources(newSpace, resources, "ingress");
+            setTimeout(() => injectResources(newSpace, resources, "ingress"), 0);
         }
         if (networkIPResourceAllowed(application)) {
             const newSpace = createSpaceForLoadedResources(networks, resources, "network");
-            injectResources(newSpace, resources, "network");
+            setTimeout(() => injectResources(newSpace, resources, "network"), 0);
         }
 
         folders.setErrors({});
@@ -381,7 +383,7 @@ export const Create: React.FunctionComponent = () => {
 
     const reloadCount = 3;
     const onLoadParameters = useCallback((importedJob: Partial<JobSpecification>) => {
-        setReloadHack({ importFrom: importedJob, count: reloadCount });
+        setReloadHack({importFrom: importedJob, count: reloadCount});
     }, []);
 
     useEffect(() => {
@@ -389,7 +391,7 @@ export const Create: React.FunctionComponent = () => {
             doLoadParameters(reloadHack.importFrom, reloadHack.count === reloadCount);
             const newCount = reloadHack.count - 1;
             if (newCount > 0) {
-                setReloadHack({ importFrom: reloadHack.importFrom, count: newCount });
+                setReloadHack({importFrom: reloadHack.importFrom, count: newCount});
             }
         }
     }, [onLoadParameters, reloadHack]);
@@ -481,16 +483,21 @@ export const Create: React.FunctionComponent = () => {
     let mandatoryWorkflow = parameters.filter(it => !it.optional && it.type === "workflow");
     if (mandatoryWorkflow.length > 1) mandatoryWorkflow = [mandatoryWorkflow[0]];
 
+    let modulesParam = parameters.filter(it => it.type === "modules");
+    if (modulesParam.length > 0) modulesParam = [modulesParam[0]];
+
+    let readmeParams = parameters.filter(it => it.type === "readme");
+
     const mandatoryParameters = parameters.filter(it =>
-        !it.optional && it.type !== "workflow"
+        !it.optional && it.type !== "workflow" && it.type !== "modules" && it.type !== "readme"
     );
 
     const activeParameters = parameters.filter(it =>
-        it.optional && activeOptParams.indexOf(it.name) !== -1
+        it.optional && activeOptParams.indexOf(it.name) !== -1 && it.type !== "readme"
     )
 
     const inactiveParameters = parameters.filter(it =>
-        !(!it.optional || activeOptParams.indexOf(it.name) !== -1)
+        !(!it.optional || activeOptParams.indexOf(it.name) !== -1) && it.type !== "readme"
     );
 
     const isMissingConnection = hasFeature(Feature.PROVIDER_CONNECTION) && estimatedCost.product != null &&
@@ -569,9 +576,9 @@ export const Create: React.FunctionComponent = () => {
                                     <div>
                                         <Flex>
                                             <ImportParameters application={application} onImport={onLoadParameters}
-                                                              importDialogOpen={importDialogOpen}
-                                                              setImportDialogOpen={setImportDialogOpen}
-                                                              onImportDialogClose={() => setImportDialogOpen(false)} />
+                                                importDialogOpen={importDialogOpen}
+                                                setImportDialogOpen={setImportDialogOpen}
+                                                onImportDialogClose={() => setImportDialogOpen(false)} />
 
                                             {anyError ?
                                                 <Tooltip trigger={
@@ -600,42 +607,42 @@ export const Create: React.FunctionComponent = () => {
                                         <div className={EstimatesContainerClass}>
                                             <table>
                                                 <tbody>
-                                                <tr>
-                                                    <th>Estimated cost</th>
-                                                    <td>
-                                                        {!estimatedCost.product ?
-                                                            "-" :
-                                                            priceToString(
-                                                                estimatedCost.product,
-                                                                estimatedCost.numberOfNodes,
-                                                                estimatedCost.durationInMinutes,
-                                                                {showSuffix: false}
-                                                            )
-                                                        }
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <th>Current balance</th>
-                                                    <td>
-                                                        {displayWallet === null ?
-                                                            "-" :
-                                                            displayWallet.usageAndQuota.display.currentBalance
-                                                        }
-                                                    </td>
-                                                </tr>
-                                                {displayWallet === null || !displayWallet.usageAndQuota.display.displayOverallocationWarning ? null :
                                                     <tr>
-                                                        <th>Usable balance</th>
+                                                        <th>Estimated cost</th>
                                                         <td>
-                                                            <OverallocationLink>
-                                                                <TooltipV2 tooltip={UNABLE_TO_USE_FULL_ALLOC_MESSAGE}>
-                                                                    <Icon name={"heroExclamationTriangle"} color={"warningMain"}/>
-                                                                    {displayWallet.usageAndQuota.display.maxUsableBalance}
-                                                                </TooltipV2>
-                                                            </OverallocationLink>
+                                                            {!estimatedCost.product ?
+                                                                "-" :
+                                                                priceToString(
+                                                                    estimatedCost.product,
+                                                                    estimatedCost.numberOfNodes,
+                                                                    estimatedCost.durationInMinutes,
+                                                                    {showSuffix: false}
+                                                                )
+                                                            }
                                                         </td>
                                                     </tr>
-                                                }
+                                                    <tr>
+                                                        <th>Current balance</th>
+                                                        <td>
+                                                            {displayWallet === null ?
+                                                                "-" :
+                                                                displayWallet.usageAndQuota.display.currentBalance
+                                                            }
+                                                        </td>
+                                                    </tr>
+                                                    {displayWallet === null || !displayWallet.usageAndQuota.display.displayOverallocationWarning ? null :
+                                                        <tr>
+                                                            <th>Usable balance</th>
+                                                            <td>
+                                                                <OverallocationLink>
+                                                                    <TooltipV2 tooltip={UNABLE_TO_USE_FULL_ALLOC_MESSAGE}>
+                                                                        <Icon name={"heroExclamationTriangle"} color={"warningMain"} />
+                                                                        {displayWallet.usageAndQuota.display.maxUsableBalance}
+                                                                    </TooltipV2>
+                                                                </OverallocationLink>
+                                                            </td>
+                                                        </tr>
+                                                    }
                                                 </tbody>
                                             </table>
                                         </div>
@@ -665,6 +672,39 @@ export const Create: React.FunctionComponent = () => {
                                 <Grid gridTemplateColumns={"1fr"} gap={"16px"} mt={"16px"}>
                                     {mandatoryWorkflow.map(param => (
                                         <Widget key={param.name} parameter={param} errors={errors} provider={provider}
+                                            injectWorkflowParameters={setWorkflowInjectParameters}
+                                            setErrors={setErrors} active application={application} />
+                                    ))}
+                                </Grid>
+                            </Card>
+                        )}
+
+                        {/*Readme*/}
+                        {readmeParams.length === 0 ? null : (
+                            <Card backgroundColor={"var(--warningMain)"} color={"warningContrast"}>
+                                <Heading.h4>
+                                    {estimatedCost.product == null ?
+                                        "Information" :
+                                        `Information from ${getProviderTitle(estimatedCost.product.category.provider)}`
+                                    }
+                                </Heading.h4>
+                                <Grid gridTemplateColumns={"1fr"} gap={"16px"} mt={"16px"}>
+                                    {readmeParams.map(param => (
+                                        <Widget key={param.name} parameter={param} errors={errors} provider={provider}
+                                                injectWorkflowParameters={setWorkflowInjectParameters}
+                                                setErrors={setErrors} active application={application} />
+                                    ))}
+                                </Grid>
+                            </Card>
+                        )}
+
+                        {/*Modules*/}
+                        {modulesParam.length === 0 ? null : (
+                            <Card>
+                                <Heading.h4>Modules</Heading.h4>
+                                <Grid gridTemplateColumns={"1fr"} gap={"16px"} mt={"16px"}>
+                                    {modulesParam.map(param => (
+                                        <Widget key={param.name} parameter={param} errors={errors} provider={provider}
                                                 injectWorkflowParameters={setWorkflowInjectParameters}
                                                 setErrors={setErrors} active application={application} />
                                     ))}
@@ -679,8 +719,8 @@ export const Create: React.FunctionComponent = () => {
                                 <Grid gridTemplateColumns={"1fr"} gap={"16px"} mt={"16px"}>
                                     {mandatoryParameters.map(param => (
                                         <Widget key={param.name} parameter={param} errors={errors} provider={provider}
-                                                injectWorkflowParameters={setWorkflowInjectParameters}
-                                                setErrors={setErrors} active application={application} />
+                                            injectWorkflowParameters={setWorkflowInjectParameters}
+                                            setErrors={setErrors} active application={application} />
                                     ))}
                                 </Grid>
                             </Card>
@@ -712,13 +752,13 @@ export const Create: React.FunctionComponent = () => {
                             <Card>
                                 <OptionalWidgetSearch pool={inactiveParameters} mapper={param => (
                                     <Widget key={param.name} parameter={param} errors={errors} provider={provider}
-                                            setErrors={setErrors}
-                                            active={false}
-                                            application={application}
-                                            onActivate={() => {
-                                                setActiveOptParams([...activeOptParams, param.name]);
-                                            }}
-                                            injectWorkflowParameters={setWorkflowInjectParameters}
+                                        setErrors={setErrors}
+                                        active={false}
+                                        application={application}
+                                        onActivate={() => {
+                                            setActiveOptParams([...activeOptParams, param.name]);
+                                        }}
+                                        injectWorkflowParameters={setWorkflowInjectParameters}
                                     />
                                 )} />
                             </Card>
@@ -726,7 +766,7 @@ export const Create: React.FunctionComponent = () => {
 
                         {/* SSH */}
                         <SshWidget application={application} onSshStatusChanged={setSshEnabled}
-                                   onSshKeysValid={setSshValid} initialEnabledStatus={initialSshEnabled} />
+                            onSshKeysValid={setSshValid} initialEnabledStatus={initialSshEnabled} />
 
                         {/* Resources */}