+++ wiki/release-notes.md
@@ -4,11 +4,39 @@ This page tracks a high-level overview of changes to the version of UCloud
 hosted at cloud.sdu.dk. This page is intended for internal
 use.
 
+## 31-01-25
+
+```
+UCloud/Core @ 2025.2.17
+UCloud/Frontend @ 2025.2.17
+IM2/Hippo @ 2025.2.17
+```
+
+## 30-01-25
+
+```
+UCloud/Core @ 2025.2.16
+```
+
+## 29-01-25
+
+```
+UCloud/Frontend @ 2025.2.13
+```
+
+## 28-01-25
+
+```
+UCloud/Core @ 2025.2.10
+UCloud/Frontend @ 2025.2.11
+```
+
 ## 27-01-25
 
 ```
 UCloud/Core @ 2025.2.9
 UCloud/Frontend @ 2025.2.9
+IM2/Hippo @ 2025.2.0
 ```